feat: only admins can add, edit, delete projects

+++ client/components/DeleteProjectModal.js
@@ -1,5 +1,6 @@
-import React from 'react'
+import React, {useContext} from 'react'
 import Modal from './Modal'
+import {AuthContext} from '../context/authContext'
 import {deleteProjectDb, getOrgDb} from '../context/axiosService'
 import {notify} from './helper/toast'
 
@@ -11,9 +12,14 @@ const DeleteProjectModal = ({
   setProjects,
   closeModal,
 }) => {
+  const {user} = useContext(AuthContext)
+
   const deleteProject = async () => {
     // close the modal BEFORE changing state
     closeModal()
+
+    if (user.status !== 'admin')
+      return notify('Only admins can delete projects!', 'warning')
 
     // delete project from db
     await deleteProjectDb(project.id)
@@ -31,8 +37,9 @@ const DeleteProjectModal = ({
     <Modal>
       <div className={styles.modalContent}>
         <div className={styles.deleteMessage}>
-          <strong>Warning!</strong> This action will delete the "{project.name}"
-          project and <strong>all columns and tasks associated with it.</strong>
+          <strong>Warning!</strong> This action will delete the project "
+          {project.name}" and{' '}
+          <strong>all columns and tasks associated with it.</strong>
           <br /> Press <span>Delete project</span> to continue, or cancel to go
           back.
         </div>
